+++ scripts/module.js
@@ -53,7 +53,7 @@ Hooks.once('init', async function() {
     setConfig();
     if (game.settings.get('chris-premades', 'Display Sidebar Macros')) enableMacroSidebar();
     if (game.settings.get('chris-premades', 'Select Tool')) enableSelectTool();
-    effectInterface.enable();
+    if (game.settings.get('chris-premades', 'Active Effect Interface')) effectInterface.enable();
 });
 Hooks.once('socketlib.ready', async function() {
     socket = socketlib.registerModule('chris-premades');
@@ -263,7 +263,7 @@ Hooks.once('ready', async function() {
     CONFIG.chrisPremades.itemConfiguration['Wild Shape'].select.compendium.values = game.packs.contents.filter(i => i.metadata.type === 'Actor').map(j => ({'value': j.metadata.id, 'html': j.metadata.label}));
     CONFIG.chrisPremades.itemConfiguration['Wild Shape'].select.compendium.default = game.settings.get('chris-premades', 'Monster Compendium');
     if (game.settings.get('chris-premades', 'Spotlight Omnisearch Summons')) Hooks.on('spotlightOmnisearch.indexBuilt', registerSearchTerms);
-    await effectInterface.startup();
+    if (game.settings.get('chris-premades', 'Active Effect Interface')) await effectInterface.startup();
 });
 let dev = {
     'setCompendiumItemInfo': setCompendiumItemInfo,